+++ src/modules/aix/server/dispatch/chatGenerate/parsers/openai.parser.ts
@@ -1,5 +1,5 @@
 import { safeErrorString } from '~/server/wire';
-import { serverSideId } from '~/server/api/trpc.nanoid';
+import { serverSideId } from '~/server/trpc/trpc.nanoid';
 
 import type { AixWire_Particles } from '../../../api/aix.wiretypes';
 import type { ChatGenerateParseFunction } from '../chatGenerate.dispatch';
